refactor: added comments
